+++ src/retrievers/getStores.ts
@@ -1,49 +1,76 @@
-// const objectAssign = require("object-assign");
-// const Store = require("../models/Store");
-// const Pagination = require("../models/Pagination");
-// const request = require("../util/request");
 import Store from "../models/Store";
 import Pagination from "../models/Pagination";
 import request from "../util/request";
-import s from "../stream";
+import stream from "../stream";
 
 const defaults = {
-  currentPage: 1,
-  pageSize: 10,
   latitude: 59.9126054,
   longitude: 10.7515334,
 };
 
-interface getStoresOptions {
+interface ISearchStoresOptions {
   currentPage?: number;
+  query?: string;
   nearLocation?: { lat: number; lon: number };
-  pageSize: number;
+  pageSize?: number;
 }
 
-interface StoreQuery {
+interface IStoreQuery {
   currentPage: number;
-  longitude?: number;
-  latitude?: number;
+  longitude: number;
+  latitude: number;
   pageSize: number;
   fields: string;
 }
 
-/**
- * @deprecated The api will no longer return all stores, but instead a max number of stores capped at 126 from the given coordinates.
- */
-export function getStores(opts?: getStoresOptions): Promise<any> {
-  const options = { ...defaults, ...opts };
-  const query: StoreQuery = {
-    fields: "BASIC",
-    currentPage: Math.max(0, options.currentPage - 1),
-    pageSize: options.pageSize,
+interface IStoreAutocompleteQuery {
+  query: string;
+}
+
+interface ISearchStoreResult {
+  stores: Store[];
+  pagination?: Pagination;
+}
+
+export function searchStores(
+  opts?: ISearchStoresOptions
+): Promise<ISearchStoreResult> {
+  if (opts?.query) {
+    return searchByQuery(opts.query);
+  }
+  const lat = opts?.nearLocation?.lat ?? defaults.latitude;
+  const lon = opts?.nearLocation?.lon ?? defaults.longitude;
+  return searchByLocation(lat, lon);
+}
+
+function searchByQuery(querystring: string): Promise<ISearchStoreResult> {
+  const query: IStoreAutocompleteQuery = {
+    query: querystring,
   };
 
-  const { lat, lon } = options.nearLocation || {};
-  if (lat && lon) {
-    query.latitude = lat;
-    query.longitude = lon;
-  }
+  const req = request.get("/api/stores/autocomplete", {
+    baseUrl: "https://www.vinmonopolet.no",
+    query,
+  });
+
+  return req.then((res) => ({
+    stores: (res.stores || []).map((i) => new Store(i)) as Store[],
+  }));
+}
+
+function searchByLocation(
+  lat: number,
+  lon: number,
+  currentPage = 1,
+  pageSize = 10
+): Promise<ISearchStoreResult> {
+  const query: IStoreQuery = {
+    fields: "BASIC",
+    currentPage: Math.max(0, currentPage - 1),
+    pageSize,
+    latitude: lat,
+    longitude: lon,
+  };
 
   const req = request.get("/api/stores", {
     baseUrl: "https://www.vinmonopolet.no",
@@ -51,31 +78,43 @@ export function getStores(opts?: getStoresOptions): Promise<any> {
   });
 
   return req.then((res) => ({
-    stores: (res.data || []).map((i) => new Store(i)),
+    stores: (res.stores || []).map((i) => new Store(i)) as Store[],
     pagination: new Pagination(
-      getPagination(query.currentPage, res),
-      options,
-      getStores
+      getPagination(query.currentPage, pageSize, res.pagination),
+      { nearLocation: { lat, lon } } as ISearchStoresOptions,
+      searchStores
     ),
   }));
 }
 
-function getPagination(currentPage, res) {
-  const pageSize = 10; // Hard coded in the API
-
+function getPagination(currentPage, pageSize, res) {
   return {
     currentPage,
     pageSize,
-    totalPages: Math.ceil(res.total / pageSize),
-    totalResults: res.total,
+    totalPages: res.totalPages,
+    totalResults: res.totalResults,
   };
 }
 
-async function getAllStores() {
-  const stream = await s.getStores();
-  stream.on("data", (res) => {
-    console.log(res);
+async function getAllStores(): Promise<Store[]> {
+  const readableStream = stream.getStores();
+  const storeStream: Promise<Store[]> = new Promise((resolve, reject) => {
+    const completeResponse: Store[] = [];
+    const onData = (chunk: Store) => {
+      // console.log(chunk);
+      completeResponse.push(chunk);
+    };
+    readableStream
+      .on("data", onData)
+      .on("end", () => {
+        resolve(completeResponse);
+      })
+      .on("error", (err: Error) => {
+        reject(err);
+      });
   });
+
+  return storeStream;
 }
 
 export default getAllStores;